+++ src/settings.js
@@ -1,10 +1,15 @@
 /**
  * Default settings. These are passed in to the main bml rendering function.
  */
-let defaultSettings = {
-  renderMarkdown: false,
+const defaultBMLSettings = {
   markdownSettings: {},
   version: null,
+};
+
+const defaultRenderSettings = {
+  randomSeed: null,
+  allowEval: true,
+  renderMarkdown: false,
   whitespaceCleanup: true,
 };
 
@@ -12,11 +17,16 @@ let defaultSettings = {
  * Return a new settings object with all the properties defined in newSettings,
  * defaulting to those in originalSettings where absent.
  *
+ * If `newSettings` is falsy, return `originalSettings` unmodified.
+ *
  * @param {Object} originalSettings
  * @param {Object} newSettings
- * @return {void}
+ * @return {Object}
  */
 function mergeSettings(originalSettings, newSettings) {
+  if (!newSettings) {
+    return originalSettings;
+  }
   let merged = JSON.parse(JSON.stringify(originalSettings));
   Object.keys(newSettings).forEach(function(key, index) {
     if (merged.hasOwnProperty(key)) {
@@ -26,5 +36,6 @@ function mergeSettings(originalSettings, newSettings) {
   return merged;
 }
 
-exports.defaultSettings = defaultSettings;
+exports.defaultBMLSettings = defaultBMLSettings;
+exports.defaultRenderSettings = defaultRenderSettings;
 exports.mergeSettings = mergeSettings;